requirements.txt
django
requests
gitpython
python-dotenv
google-genai
gunicorn
mysqlclient
psycopg2